restructure frontend file structure
