+++ gatsby-config.js
@@ -41,7 +41,7 @@ module.exports = {
       resolve: 'gatsby-source-filesystem',
       options: {
         name: 'images',
-        path: 'images',
+        path: 'static/images',
       },
     },
   ],
