+++ scripts/fix_image_links.py
